requirements.txt
pygame
pyserial
RPi.GPIO
pydub
python-slugify
